Added CRFitAuto and CRInterp.
Added CRFitAuto and CRFitAutoTest, which automatically sets a sufficient number of terms to fit for the fitted complex rational polygonal and sets sufficient numerical precision for an accurate result.

Added CRInterp that uses CRFitAuto to fit a complex rational polynomial to each S/Y/Z parameter in an SnP data structure. Significant extrapolation also often successful.

Modfied SnP.Shorten to copy the SnP stucture to be shortened to include only data from frequencies whose indices are passed to it and return the shortened copy of the passed SnP.

Modified SnP.Fill to allow filling of S.Y.Z parameters, not jus Y parameters.

Modified SnP constructor to fill the created SnP structure with data from an SnP file when the file name is paaed.